+++ Pinetree.Client/Pages/Edit.razor
@@ -1,13 +1,13 @@
 ﻿@page "/Edit/{id:long}"
 @attribute [Authorize]
 
-@using Microsoft.AspNetCore.Authorization
-@using Pinetree.Client.Model
-@using Pinetree.Client.Services
-@using Pinetree.Shared
-@using Pinetree.Shared.Model
 @using System.Net.Http
 @using System.Security.Claims
+@using Microsoft.AspNetCore.Authorization
+@using Pinetree.Client.Services
+@using Pinetree.Client.VModel
+@using Pinetree.Shared
+@using Pinetree.Shared.Model
 
 @rendermode InteractiveWebAssembly
