+++ Pinetree.Client/Pages/List.razor
@@ -2,7 +2,7 @@
 @attribute [Authorize]
 
 @using Microsoft.AspNetCore.Authorization
-@using Pinetree.Client.Model
+@using Pinetree.Client.VModel
 @using Pinetree.Client.Services
 @using Pinetree.Shared.Model
 @using System.Net.Http
