+++ Pinetree.Client/Pages/Tryit.razor
@@ -1,6 +1,6 @@
 ﻿@page "/Tryit"
 
-@using Pinetree.Client.Model
+@using Pinetree.Client.VModel
 
 @rendermode InteractiveAuto
